Experimental: added 64-bit pos, mate pos and insert size to CRAM V4.0.
The BAM internal layout is 64-bit too, and SAM can fill this out with
64-bit fields.  This permits SAM to CRAM to SAM round trips with long
references, although BAM is still unable to handle this (and it
doesn't report this failure yet).

+++ io_lib/cram_structs.h
@@ -276,8 +276,8 @@ struct cram_map;
 /* Compression header block */
 typedef struct {
     int32_t ref_seq_id;
-    int32_t ref_seq_start;
-    int32_t ref_seq_span;
+    int64_t ref_seq_start;
+    int64_t ref_seq_span;
     int32_t num_records;
     int32_t num_landmarks;
     int32_t *landmark;
@@ -328,8 +328,8 @@ typedef struct {
 typedef struct {
     enum cram_content_type content_type;
     int32_t ref_seq_id;     /* if content_type == MAPPED_SLICE */
-    int32_t ref_seq_start;  /* if content_type == MAPPED_SLICE */
-    int32_t ref_seq_span;   /* if content_type == MAPPED_SLICE */
+    int64_t ref_seq_start;  /* if content_type == MAPPED_SLICE */
+    int64_t ref_seq_span;   /* if content_type == MAPPED_SLICE */
     int32_t num_records;
     int64_t record_counter;
     int32_t num_blocks;
@@ -356,8 +356,8 @@ struct ref_entry;
 typedef struct {
     int32_t  length;
     int32_t  ref_seq_id;
-    int32_t  ref_seq_start;
-    int32_t  ref_seq_span;
+    int64_t  ref_seq_start;
+    int64_t  ref_seq_span;
     int64_t  record_counter;
     int64_t  num_bases;
     int32_t  num_records;
@@ -379,16 +379,16 @@ typedef struct {
     int max_c_rec, curr_c_rec;   // current and max recs per container
     int slice_rec;               // rec no. for start of this slice
     int curr_ref;                // current ref ID. -2 for no previous
-    int last_pos;                // last record position
+    int64_t last_pos;                // last record position
     struct cram_slice **slices, *slice;
     int pos_sorted;              // boolean, 1=>position sorted data
-    int max_apos;                // maximum position, used if pos_sorted==0
+    int64_t max_apos;                // maximum position, used if pos_sorted==0
     int last_slice;              // number of reads in last slice (0 for 1st)
     int multi_seq;               // true if packing multi seqs per cont/slice
     int unsorted;		 // true is AP_delta is 0.
 
     /* Copied from fd before encoding, to allow multi-threading */
-    int ref_start, first_base, last_base, ref_id, ref_end;
+    int64_t ref_start, first_base, last_base, ref_id, ref_end;
     char *ref;
     //struct ref_entry *ref;
 
@@ -420,14 +420,14 @@ typedef struct {
     int32_t flags;        // BF
     int32_t cram_flags;   // CF
     int32_t len;          // RL
-    int32_t apos;         // AP
+    int64_t apos;         // AP
     int32_t rg;           // RG
     int32_t name;         // RN; idx to s->names_blk
     int32_t name_len;
     int32_t mate_line;    // index to another cram_record
     int32_t mate_ref_id;
-    int32_t mate_pos;     // NP
-    int32_t tlen;         // TS
+    int64_t mate_pos;     // NP
+    int64_t tlen;         // TS
 
     // Auxiliary data
     int32_t ntags;        // TC
@@ -444,7 +444,7 @@ typedef struct {
     int32_t qual;         // idx to s->qual_blk
     int32_t cigar;        // idx to s->cigar
     int32_t ncigar;
-    int32_t aend;         // alignment end
+    int64_t aend;         // alignment end
     int32_t mqual;        // MQ
 
     int32_t feature;      // idx to s->feature
@@ -546,7 +546,7 @@ typedef struct cram_slice {
     cram_block **block_by_id;
 
     /* State used during encoding/decoding */
-    int last_apos, max_apos;
+    int64_t last_apos, max_apos;
 
     /* Array of decoded cram records */
     cram_record *crecs;
@@ -669,8 +669,8 @@ typedef struct cram_index {
 
 typedef struct {
     int refid;
-    int start;
-    int end;
+    int64_t start;
+    int64_t end;
 } cram_range;
 
 /*-----------------------------------------------------------------------------
@@ -778,7 +778,7 @@ typedef struct {
     cram_container *ctr_mt;
 
     // positions for encoding or decoding
-    int first_base, last_base;
+    int64_t first_base, last_base;
 
     // cached reference portion
     refs_t *refs;              // ref meta-data structure
